Refactored to allow custom names. Also, modified MAC editing to not change words that end in a, c, i, o, j or z. Changed the apostrophe editing to handle a space after the apostrophe.
Still need to handle names like "Dan Jones' mother".
