+++ src/NameCase/NameCase/FormattedNameFactory.cs
@@ -1,0 +1,75 @@
+﻿namespace PgbNameCase;
+
+public class FormattedNameFactory
+{
+	private static readonly Lazy<HashSet<string>> _Names = new(() => new HashSet<string>
+	{
+			"AJ",
+			"BJ",
+			"DeRosa" ,
+			"DeWitt" ,
+			"JD",
+			"JJ",
+			"Macaraeg",
+			"Maccarone",
+			"Macchione",
+			"Macevicius",
+			"Machacek",
+			"Machak",
+			"Machamer",
+			"Machan",
+			"Machar",
+			"Machart",
+			"Machell",
+			"Machemer",
+			"Machen",
+			"Macher",
+			"Machiel",
+			"Machin",
+			"Machlin",
+			"Machnik",
+			"Machon",
+			"Maciag",
+			"Macias",
+			"Maciel",
+			"Macioce",
+			"Maciol",
+			"Maciolek",
+			"Maciulis",
+			"Mack",
+			"Mackall",
+			"Mackel",
+			"Mackell",
+			"Macken",
+			"Mackenthun",
+			"Macker",
+			"Mackert",
+			"Mackes",
+			"Mackey",
+			"Mackie",
+			"Mackin",
+			"Mackintosh",
+			"Mackle",
+			"Macklem",
+			"Mackler",
+			"Mackley",
+			"Macklin",
+			"Mackowiak",
+			"Mackrell",
+			"Maclin",
+			"Macomb",
+			"Macomber",
+			"Macquarie",
+			"Macris",
+			"Macumber",
+			"LaRosa" ,
+			"LaRue",
+
+			// Roman numerals
+			"II", 
+			"III", 
+			"IV"
+	}, true); // `true` enables thread safety
+
+	public static IReadOnlySet<string> GetNames() => _Names.Value;
+}
