+++ src/NameCase/NameCase/StartsWithFormatter.cs
@@ -5,18 +5,14 @@ internal class StartsWithFormatter
 	internal static string Format(string nameToBeFormatted)
 	{
 		var position = 0;
-		var specials = new List<string>();
 
-		if (nameToBeFormatted.Length > 1 && nameToBeFormatted.Substring(1, 1) == "'")
-		{
-			position = 2;
-		}
 
 		if (nameToBeFormatted.StartsWith("Mac"))
 		{
 			if (nameToBeFormatted.Length <= 5) return nameToBeFormatted;
 
-			specials = SpecialHandlingFactory.GetMacSpecials();
+			if (nameToBeFormatted.EndsWithSpecifiedChars()) return nameToBeFormatted;
+
 			position = 3;
 		}
 
@@ -24,7 +20,8 @@ internal class StartsWithFormatter
 		{
 			if (nameToBeFormatted.Length <= 4) return nameToBeFormatted;
 
-			specials = SpecialHandlingFactory.GetMcSpecials();
+			if (nameToBeFormatted.EndsWithSpecifiedChars()) return nameToBeFormatted;
+
 			position = 2;
 		}
 
@@ -36,19 +33,18 @@ internal class StartsWithFormatter
 		if (nameToBeFormatted.StartsWith("Mc'"))
 		{
 			if (nameToBeFormatted.Length <= 5) return nameToBeFormatted;
+
 			position = 3;
 		}
 
 		if (position > 0)
 		{
-			if (!specials.Contains(nameToBeFormatted))
-			{
-				nameToBeFormatted = UpperSpecificPosition(nameToBeFormatted, position);
-			}
+			nameToBeFormatted = UpperSpecificPosition(nameToBeFormatted, position);
 		}
 
 		return nameToBeFormatted;
 	}
+
 
 	private static string UpperSpecificPosition(string nameToBeFormatted, int position)
 	{
