+++ src/NameCase/UnitTests/NameCaseFormatterShould/GivenANameWithAParticle.cs
@@ -1,6 +1,6 @@
 ﻿using PgbNameCase;
 
-namespace UnitTests;
+namespace UnitTests.NameCaseFormatterShould;
 
 public class GivenANameWithAParticle
 {
