+++ src/NameCase/UnitTests/NameCaseFormatterShould/GivenATwoLetterName.cs
